prov/efa: move shm_info and shm_name functions to efa_shm.c
move shm related functions to efa_shm.c and renamed them
accordingly.
