+++ prov/efa/Makefile.include
@@ -33,6 +33,7 @@ if HAVE_EFA
 _efa_files = \
 	prov/efa/src/efa_device.c \
 	prov/efa/src/efa_hmem.c \
+	prov/efa/src/efa_shm.c \
 	prov/efa/src/efa_av.c \
 	prov/efa/src/efa_domain.c \
 	prov/efa/src/efa_cm.c \
@@ -64,6 +65,7 @@ _efa_files = \
 _efa_headers = \
 	prov/efa/src/efa.h \
 	prov/efa/src/efa_mr.h \
+	prov/efa/src/efa_shm.h \
 	prov/efa/src/efa_hmem.h \
 	prov/efa/src/efa_device.h \
 	prov/efa/src/efa_domain.h \
