+++ prov/efa/src/rxr/rxr_atomic.c
@@ -48,7 +48,7 @@ static void rxr_atomic_init_shm_msg(struct fi_msg_atomic *shm_msg,
 
 	assert(msg->rma_iov_count <= RXR_IOV_LIMIT);
 	memcpy(shm_msg, msg, sizeof(*msg));
-	if (!(shm_info->domain_attr->mr_mode & FI_MR_VIRT_ADDR)) {
+	if (!(g_shm_info->domain_attr->mr_mode & FI_MR_VIRT_ADDR)) {
 		memcpy(rma_iov, msg->rma_iov,
 		       sizeof(*msg->rma_iov) * msg->rma_iov_count);
 		for (i = 0; i < msg->rma_iov_count; i++)
@@ -242,7 +242,7 @@ rxr_atomic_inject(struct fid_ep *ep,
 	peer = rxr_ep_get_peer(rxr_ep, dest_addr);
 	assert(peer);
 	if (peer->is_local && rxr_ep->use_shm_for_tx) {
-		if (!(shm_info->domain_attr->mr_mode & FI_MR_VIRT_ADDR))
+		if (!(g_shm_info->domain_attr->mr_mode & FI_MR_VIRT_ADDR))
 			remote_addr = 0;
 
 		return fi_inject_atomic(rxr_ep->shm_ep, buf, count, peer->shm_fiaddr,
